migration change

+++ database/migrations/2022_10_16_201955_create_factures_table.php
@@ -21,9 +21,9 @@ class CreateFacturesTable extends Migration
             $table->string('mode_paiement');
             $table->string('notes');
             $table->float("totalFacture");
-            $table->integer("prestation_id");
-            $table->integer("product_id");
-            $table->integer("client_id");
+            $table->unsignedBigInteger("prestation_id");
+            $table->unsignedBigInteger("product_id");
+            $table->unsignedBigInteger("client_id");
             $table->timestamps();
         });
     }
